Small refactor of Bootstrap.kt to make the file a bit more readable
