+++ src/main/kotlin/no/nav/syfo/client/SamhandlerPraksisType.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo
+package no.nav.syfo.client
 
 enum class SamhandlerPraksisType(val kodeVerdi: String) {
     AUDIOPEDAGOG("AUDI"),
