+++ src/test/kotlin/no/nav/syfo/ArsakKodeMissingSpek.kt
@@ -2,7 +2,10 @@ package no.nav.syfo
 
 import java.io.StringReader
 import no.nav.helse.sm2013.HelseOpplysningerArbeidsuforhet
+import no.nav.syfo.util.aktivitetIkkeMuligMissingArbeidsplassenArsakskode
+import no.nav.syfo.util.aktivitetIkkeMuligMissingMedisinskeArsakskode
 import no.nav.syfo.util.fellesformatUnmarshaller
+import no.nav.syfo.util.medisinskeArsakskodeIsmissing
 import no.nav.syfo.utils.getFileAsString
 import org.amshove.kluent.shouldEqual
 import org.spekframework.spek2.Spek
@@ -11,7 +14,7 @@ import org.spekframework.spek2.style.specification.describe
 object ArsakKodeMissingSpek : Spek({
 
     describe("Validate arsakskodeIsmissing is false") {
-        it("Validate MedisinskeArsaker Arsakskode is mapped") {
+        it("Validate Medisinske qArsaker Arsakskode is mapped") {
             val healthInformation = fellesformatUnmarshaller.unmarshal(StringReader(getFileAsString("src/test/resources/generated_sm_8.xml"))) as HelseOpplysningerArbeidsuforhet
 
             medisinskeArsakskodeIsmissing(healthInformation) shouldEqual false
