+++ src/test/kotlin/no/nav/syfo/BehandlerMissingFNRDNRSpek.kt
@@ -3,6 +3,7 @@ package no.nav.syfo
 import java.io.StringReader
 import no.nav.helse.sm2013.HelseOpplysningerArbeidsuforhet
 import no.nav.syfo.util.fellesformatUnmarshaller
+import no.nav.syfo.util.fnrAndDnrIsmissingFromBehandler
 import no.nav.syfo.utils.getFileAsString
 import org.amshove.kluent.shouldEqual
 import org.spekframework.spek2.Spek
